move Queue to a separate file and add a builder for when future options are added
